requirements.txt
python-dotenv==1.1.0
requests>=2.28.0
groq==0.23.1
PyPDF2>=3.0.0
